#291: refactored Commit class into Commit.js
